Move `Evcxr`, `Grcov` and `WasmPack` to the `binaries` package
