+++ src/main/kotlin/org/rust/cargo/toolchain/binaries/WasmPack.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.cargo.toolchain
+package org.rust.cargo.toolchain.binaries
 
 import com.intellij.execution.configurations.GeneralCommandLine
 import org.rust.cargo.util.splitOnDoubleDash
